+++ providers/kmpauth-google/src/jsMain/kotlin/com/mmk/kmpauth/google/GoogleAuthUiProviderImpl.kt
@@ -24,9 +24,13 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
         filterByAuthorizedAccounts: Boolean,
         isAutoSelectEnabled: Boolean,
         scopes: List<String>
-    ): GoogleUser? {
+    ): Result<GoogleUser> {
         val scriptLoaded = waitForGoogleAuthScriptToLoad()
-        if (!scriptLoaded) return null
+        if (!scriptLoaded) {
+            return Result.failure(
+                IllegalStateException("Google Sign-In script failed to load")
+            )
+        }
 
         return suspendCancellableCoroutine { continuation ->
             val tokenClientConfig = createTokenClientConfig(
@@ -56,12 +60,12 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
         document.head?.appendChild(script)
     }
 
-    private suspend fun CancellableContinuation<GoogleUser?>.handleTokenResponse(tokenResponse: dynamic) {
+    private suspend fun CancellableContinuation<Result<GoogleUser>>.handleTokenResponse(tokenResponse: dynamic) {
 
         val error = getTokenResponseError(tokenResponse)
         if (error != null) {
             showConsoleError("Error during Google sign-in: $error")
-            resume(null)
+            resume(Result.failure(IllegalStateException("Google sign-in failed: $error")))
             return
         }
 
@@ -83,7 +87,7 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
                 profilePicUrl = picture
             )
 
-            resume(googleUser)
+            resume(Result.success(googleUser))
 
         } catch (err: Throwable) {
             showConsoleError("Error fetching user info: $err")
@@ -92,7 +96,7 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
                 idToken = idToken,
                 accessToken = accessToken
             )
-            resume(googleUser)
+            resume(Result.success(googleUser))
         }
     }